+++ zingolib/src/testutils/macros.rs
@@ -1,50 +1,5 @@
 //! TODO
 
-/// Note that `do_addresses` returns an array, each element is a JSON representation
-/// of a UA.  Legacy addresses can be extracted from the receivers, per:
-/// <https://zips.z.cash/zip-0316>
-// TODO: is this needed as a macro?
-// TODO: change unified to orchard as both are unified with orchard-only or sapling-only receiver selections
-#[macro_export]
-macro_rules! get_base_address_macro {
-    ($client:expr, $address_protocol:expr) => {
-        match $address_protocol {
-            "unified" => {
-                assert_eq!(
-                    $client.unified_addresses_json().await[0]["has_orchard"]
-                        .as_bool()
-                        .unwrap(),
-                    true
-                );
-                $client.unified_addresses_json().await[0]["encoded_address"]
-                    .clone()
-                    .to_string()
-            }
-            "sapling" => {
-                assert_eq!(
-                    $client.unified_addresses_json().await[1]["has_orchard"]
-                        .as_bool()
-                        .unwrap(),
-                    false
-                );
-                assert_eq!(
-                    $client.unified_addresses_json().await[1]["has_sapling"]
-                        .as_bool()
-                        .unwrap(),
-                    true
-                );
-                $client.unified_addresses_json().await[1]["encoded_address"]
-                    .clone()
-                    .to_string()
-            }
-            "transparent" => $client.transparent_addresses_json().await[0]["encoded_address"]
-                .clone()
-                .to_string(),
-            _ => "ERROR".to_string(),
-        }
-    };
-}
-
 /// First check that each pools' balance matches an expectation
 /// then check that the overall balance as calculated by
 /// summing the amounts listed in `tx_summaries` matches the